refactor(ui): split config adapter schema ownership
